Updated the clock app
When pressed on the Add button on the Clock view, a row list is created for the Alarm. Now, when the user clicks on one of the rows / Alarms, they can go to a different view set to SetClockView. For now only back and save are there as toolbar buttons. I need to figure out how to save the row data when the user closes the app. i also need to add a way to set a clock time.